requirements.txt
adal==1.2.4
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
azure-batch==9.0.0
azure-common==1.1.25
azure-core==1.5.0
azure-mgmt-resource==10.2.0
azure-storage-blob==12.3.1
backcall==0.2.0
beautifulsoup4==4.9.0
bleach==3.1.5
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
common==0.1.2
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-adminrestrict==2.0.2
django-contact-form==1.8.1
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-filter==2.2.0
django-phone-field==1.8.0
django-plotly-dash==1.3.1
django-ranged-response==0.2.0
django-redis-cache==2.1.1
django-rq==2.3.2
django-simple-captcha==0.5.12
django-simple-menu==1.2.1
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
dpd-components==0.1.0
entrypoints==0.3
fake-useragent==0.1.11
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
geojson==2.5.0
git-filter-repo==2.27.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
md-mermaid==0.1.0
mistune==0.8.4
msrest==0.6.14
msrestazure==0.6.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parse==1.16.0
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
pip-tools==5.0.0
plotly==4.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
pycparser==2.20
pyee==7.0.2
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
redis==3.5.3
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.4.2
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
sendgrid==6.2.2
Shapely==1.7.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
turfpy==0.0.3
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.2.4
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wget==3.2
wheel=0.34.2
zipp==3.1.0
